requirements.txt
openai>=1.14.2
requests>=2.31.0
colorama>=0.4.6
aiohttp>=3.9.4
discord.py>=2.3.2
python-dotenv>=1.0.1